Rename catalog to catalogue
This is so we follow the decided naming

# Conflicts:
#	catalog/src/main/kotlin/com/tidal/sdk/catalog/Catalog.kt
#	catalogue/.openapi-generator-ignore
#	catalogue/.openapi-generator/FILES
#	catalogue/.openapi-generator/VERSION
#	catalogue/docs/AlbumAttributes.md
#	catalogue/docs/AlbumDataDocument.md
#	catalogue/docs/AlbumJSONAPI.md
#	catalogue/docs/AlbumJSONAPIApi.md
#	catalogue/docs/AlbumRelationship.md
#	catalogue/docs/AlbumRelationshipResource.md
#	catalogue/docs/AlbumRelationshipsDocument.md
#	catalogue/docs/AlbumResource.md
#	catalogue/docs/AlbumsDataDocument.md
#	catalogue/docs/ArtistAttributes.md
#	catalogue/docs/ArtistDataDocument.md
#	catalogue/docs/ArtistJSONAPI.md
#	catalogue/docs/ArtistJSONAPIApi.md
#	catalogue/docs/ArtistRelationship.md
#	catalogue/docs/ArtistRelationshipResource.md
#	catalogue/docs/ArtistRelationshipsDocument.md
#	catalogue/docs/ArtistResource.md
#	catalogue/docs/ArtistsDataDocument.md
#	catalogue/docs/AvailableAlbumRelationships.md
#	catalogue/docs/AvailableAlbumRelationshipsItems.md
#	catalogue/docs/AvailableAlbumRelationshipsItemsDataInner.md
#	catalogue/docs/AvailableAlbumRelationshipsItemsDataInnerMeta.md
#	catalogue/docs/AvailableArtistRelationships.md
#	catalogue/docs/AvailableTrackRelationships.md
#	catalogue/docs/AvailableVideoRelationships.md
#	catalogue/docs/ErrorDocument.md
#	catalogue/docs/ErrorObject.md
#	catalogue/docs/ErrorObjectSource.md
#	catalogue/docs/ExternalLink.md
#	catalogue/docs/ExternalLinkMeta.md
#	catalogue/docs/ImageLink.md
#	catalogue/docs/ImageLinkMeta.md
#	catalogue/docs/Links.md
#	catalogue/docs/ProviderAttributes.md
#	catalogue/docs/ProviderDataDocument.md
#	catalogue/docs/ProviderJSONAPI.md
#	catalogue/docs/ProviderJSONAPIApi.md
#	catalogue/docs/ProviderRelationshipResource.md
#	catalogue/docs/ProviderResource.md
#	catalogue/docs/ProvidersDataDocument.md
#	catalogue/docs/Relationship.md
#	catalogue/docs/ResourceIdentifier.md
#	catalogue/docs/ResourceRelationship.md
#	catalogue/docs/TrackAttributes.md
#	catalogue/docs/TrackDataDocument.md
#	catalogue/docs/TrackJSONAPI.md
#	catalogue/docs/TrackJSONAPIApi.md
#	catalogue/docs/TrackRelationship.md
#	catalogue/docs/TrackRelationshipResource.md
#	catalogue/docs/TrackRelationshipsDocument.md
#	catalogue/docs/TrackResource.md
#	catalogue/docs/TracksDataDocument.md
#	catalogue/docs/VideoAttributes.md
#	catalogue/docs/VideoDataDocument.md
#	catalogue/docs/VideoJSONAPI.md
#	catalogue/docs/VideoJSONAPIApi.md
#	catalogue/docs/VideoLink.md
#	catalogue/docs/VideoLinkMeta.md
#	catalogue/docs/VideoRelationship.md
#	catalogue/docs/VideoRelationshipResource.md
#	catalogue/docs/VideoRelationshipsDocument.md
#	catalogue/docs/VideoResource.md
#	catalogue/docs/VideosDataDocument.md
#	catalogue/gradle/wrapper/gradle-wrapper.jar
#	catalogue/gradle/wrapper/gradle-wrapper.properties
#	catalogue/gradlew
#	catalogue/gradlew.bat
#	catalogue/openapi.json
#	catalogue/proguard-rules.pro
#	catalogue/settings.gradle
