+++ catalogue/apps/demo/src/main/AndroidManifest.xml
