+++ catalogue/apps/demo/src/main/kotlin/com/tidal/sdk/demo/DemoApp.kt
