+++ catalogue/apps/demo/src/main/res/values/strings.xml
