+++ catalogue/src/main/AndroidManifest.xml
