+++ settings.gradle.kts
@@ -50,4 +50,4 @@ listOf(
 includeFromDefaultHierarchy("auth")
 includeFromDefaultHierarchy("common")
 includeFromDefaultHierarchy("eventproducer")
-includeFromDefaultHierarchy("catalog")
+includeFromDefaultHierarchy("catalogue")
